relocate rosetta_request_handler
