+++ app/lib/api.py
@@ -75,3 +75,14 @@ class JSONAPIClient:
             raise ResourceNotFound("Resource not found")
         logger.error(f"JSON API responded with {response.status_code}")
         raise Exception("Request failed")
+
+
+def rosetta_request_handler(uri, params={}) -> dict:
+    """Prepares and initiates the api url requested and returns response data"""
+    api_url = settings.ROSETTA_API_URL
+    if not api_url:
+        raise Exception("ROSETTA_API_URL not set")
+    client = JSONAPIClient(api_url)
+    client.add_parameters(params)
+    data = client.get(uri)
+    return data
